Restructure project with packages and Gradle subprojects
