+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'intro17'
 
+include 'countzeroes'; 'main'
